[Commit] 2023/05/05 2

+++ client/build.bat
@@ -5,11 +5,13 @@ set APK_PATH=%ROOT_PATH%apk
 
 call flutter build apk --target-platform android-arm64 --split-per-abi
 
-rmdir /s /q %APK_PATH%
+if exist %APK_PATH% rmdir /s /q %APK_PATH%
 mkdir %APK_PATH%
 
 cd %ROOt_PATH%build\app\outputs\flutter-apk
 
 copy app-arm64-v8a-release.apk %APK_PATH%\
+
+cd %APK_PATH%
 
 rename %APK_PATH%\app-arm64-v8a-release.apk app-release.apk
