+++ client/generate.bat
@@ -3,6 +3,6 @@
 set ROOT_PATH=%~dp0
 set GENERATE_PATH=%ROOT_PATH%lib\generate
 
-rmdir /s /q %GENERATE_PATH%
+if exist %GENERATE_PATH% rmdir /s /q %GENERATE_PATH%
 
 call flutter packages pub run build_runner build
